refactor(takum): split takum and takum_log into core / manipulators / iostream / debug (#1334)
Phase 2 continues after lns. One core covers both types, since takum.hpp has always
included both impl headers.

  header                              lines  headers   I/O
  takum/takum.hpp   (before)          71865      428     5
  takum/core.hpp    (new)             49429      301     0

CRITERION 2 ESSENTIALLY MET, and this is the interesting result. Every float-like so far
landed near 78,000 lines -- cfloat 78330, posit 77430, lns 74762 -- all pinned by
blocktriple's ~69,800. takum uses NO blocktriple: grep finds zero references in either
impl header. Freed of that floor it lands at 49,429, within 10% of the 45,000 target and
second only to integer's 48,328.

That is worth recording for the rest of the rollout: the ~78k plateau is not what a
layered core costs, it is what depending on blocktriple costs. Types built directly on
blockbinary land near 49k. It also sharpens the blocktriple question -- if the 45,000
target is ever made binding, blocktriple is the whole of the gap for the types that use
it, and nothing else needs attention.

Preceded by the #1389 sweep over the takum subtree using the hardened script: 24 gaps
across 33 headers, plus 10 more that only became visible once the text moved and its
dependencies had to be named where they landed.

Out of the core: debugConstexprParameters() to a new debug.hpp; to_binary and to_native
for BOTH types to manipulators.hpp; operator<< / operator>> for both to a new
iostream.hpp. native/ieee754.hpp became ieee754_core.hpp in both impl headers, and
takum_log.hpp now gets takum_log_impl.hpp through the core rather than directly.

The extraction of takum_log's stream operators split operator>> mid-body on the first
attempt -- my heuristic stopped at the last line matching operator(<<|>>) within a
window, and the body ran past it, leaving an orphan `double d; istr >> d; ...` behind.
322 compile errors made it obvious. Repaired, and a to_native in each impl header that
the same grep had missed was picked up in the process.

Verified: both gates pass on gcc and clang, for takum and takum_log; takum suite 13 pass
0 fail; every api.cpp 55 pass 0 fail; the hardened sweep reports 0 gaps across all 36
takum headers.
