+++ include/sw/universal/number/takum/takum_traits.hpp
@@ -5,6 +5,7 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <type_traits>   // std::enable_if_t
 #include <universal/traits/integral_constant.hpp>
 
 namespace sw { namespace universal {
